+++ packages/cloudflare/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @authhero/cloudflare-adapter
 
+## 2.7.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @authhero/adapter-interfaces@0.97.0
+
 ## 2.7.4
 
 ### Patch Changes